+++ app/Domain/News/Entities/News.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace App\Domain\News\Entities;
+
+use Illuminate\Database\Eloquent\Factories\HasFactory;
+use Illuminate\Database\Eloquent\Model;
+
+class News extends Model
+{
+    use HasFactory;
+
+    protected $fillable = [
+        'title',
+        'description',
+        'full_txt',
+        'image',
+        'publish_date'
+    ];
+
+}