+++ HISTORY.md
@@ -1,3 +1,28 @@
-## 0.1.0 (2024-02-27)
+## 1.0.5 (2024-02-27)
+
+
+
+## 1.0.4 (2024-02-27)
+
+* criando EloquentNewsRepository que faz a implementaçao da NewsRepositoryInterface
+* instanciando os metodos da interface NewsRepositoryInterface
+* 
+## 1.0.3 (2024-02-27)
+
+* criando service NewsService
+* instanciando os metodos da interface NewsRepositoryInterface
+
+## 1.0.2 (2024-02-27)
+
+* criando interface NewsRepositoryInterface
+* criando metodos da interface
+
+
+## 1.0.1 (2024-02-27)
+
+* criando model news
+* criando migrate matters
+
+## 1.0.0 (2024-02-27)
 
 * Iniciando o projeto